feat: choose bindAddress on our own by default
This now chooses the bindAddress on our own using a similar mechanism
we suggest for users to find out their bindAddress (which is the source
IP address of the default route).

The bindAddress can, however, be overwritten by the user.
