+++ docker-compose.yaml
@@ -4,7 +4,6 @@ services:
   bee:
     image: ghcr.io/cybersecurity-and-enterprise-security/bee:main
     restart: always
-    command: -bind ...
     network_mode: host
     cap_add:
       - NET_ADMIN
